[move] Button context と modifier を別ファイルに移動しました.
